remove msgpack

+++ protostuff-cpe/src/main/java/io/protostuff/test/Foo.java
@@ -3,7 +3,7 @@ package io.protostuff.test;
 public class Foo
 {
     short name;
-    int age;
+    long age;
 
     public Foo(short name, int age)
     {
